\name{rglwidget-package}
\alias{rglwidget-package}
\alias{rglwidget}
\alias{renderRglwidget}
\alias{rglwidgetOutput}
\alias{playwidget}
\alias{renderPlaywidget}
\alias{playwidgetOutput}
\alias{subsetControl}
\alias{propertyControl}
\alias{clipplaneControl}
\alias{ageControl}
\alias{vertexControl}
\alias{elementId2Prefix}
\alias{registerSceneChange}
\alias{sceneChange}
\alias{\%>\%}
\docType{package}
\title{
\packageTitle{rglwidget}
}
\description{
\packageDescription{rglwidget}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rglwidget}
\packageIndices{rglwidget}

The \pkg{rglwidget} package has been merged into the
\pkg{rgl} package, and should no longer be used.
}
\author{
\packageAuthor{rglwidget}

Maintainer: \packageMaintainer{rglwidget}
}
\keyword{ package }
